requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
Pillow==10.2.0
lxml==5.1.0